Extract helper methods to a shared module
Extract helper methods to a shared module
