+++ spec/models/manager_refresh/save_inventory/single_dto_collection_spec.rb
@@ -1,6 +1,8 @@
+require_relative 'spec_helper'
 require_relative 'spec_parsed_data'
 
 describe ManagerRefresh::SaveInventory do
+  include SpecHelper
   include SpecParsedData
 
   ######################################################################################################################
@@ -597,17 +599,5 @@ describe ManagerRefresh::SaveInventory do
         end
       end
     end
-  end
-
-  def assert_all_records_match_hashes(model_classes, *expected_match)
-    # Helper for matching attributes of the model's records to an Array of hashes
-    model_classes = model_classes.kind_of?(Array) ? model_classes : [model_classes]
-    attributes    = expected_match.first.keys
-    model_classes.each { |m| expect(m.to_a.map { |x| x.slice(*attributes) }).to(match_array(expected_match)) }
-  end
-
-  def add_data_to_dto_collection(dto_collection, *args)
-    # Creates Dto object from each arg and adds it into the DtoCollection
-    args.each { |data| dto_collection << dto_collection.new_dto(data) }
   end
 end
